+++ plugins/Tour/Engagement/ChallengeScheduledReport.php
@@ -44,10 +44,10 @@ class ChallengeScheduledReport extends Challenge
         return 'add_scheduled_report';
     }
 
-    public function isCompleted()
+    public function isCompleted(string $login)
     {
         if (!isset($this->completed)) {
-            $this->completed = $this->finder->hasAddedNewEmailReport(Piwik::getCurrentUserLogin());
+            $this->completed = $this->finder->hasAddedNewEmailReport($login);
         }
         return $this->completed;
     }